Add default severity for custom error levels
Some python project use custom (not build-in) error levels for their loggers(for example python). Currently when you use custom levels the Airbrake handler will fail constructing the airbrake error because of the level not existing as a build-in ErrorLevel. As a quick fix I changed the getattr call to return the custom error level when no build-in error level is found.